have csv_stringclean interpret characters to be nuked as plain characters.
Interpretting a string containing characters to be nuked as a QRegularExpression
is problematic if the strings means something special as a regular expression.
For example [:\n:] is not a valid QRegularExpression.
This expression is generated by xmapwpt.
For example [0-9] as a QRegularExpression means characters 0 through 9, instead
of characters 0, -, and 9.
For example [^a] as a QRegularExpression means not characater a, instaed of
characters ^ and a.

QRegExp is deprecated and has similar although not identical problems as
QRegularExpression.
For example [:\n:] is a valid QRegExp, but not a valid QRegularExpression.