Fix kml bounding box, and introduce NVectors.
The kml bouning box had problems around the antimeridian, that were
unsuccessfully mitigated in 2011. Instead of fixing the obvious
mistake in that mitigation we use nvectors to compute the actual
center of the bounding box.
Nvectors will be useful to avoid problems at discontinuities and
singularites that are inherent in the latitude,longitude coordinate
system.