add calibration support for humminbird.
This conditionally implements calibration of the humminbird
conversion constants EAST_SCALE and cos_ae.
We assume reference/humminbird.hwr was created with hand crafted
waypoints whose latitude and longitude had zero fractional part.
We use the Newton Raphson method to find the values of the
constants that minimizes the error for the conversions from
(lat,lon) -> (north,east) using this set of points.
It would be better to have many additional points to average out
the quantization noise from the integer north & east values, which
have a resolution of approximately 1 meter.
Alternatively if more precise latitude and longitude value were
available from conversion of the integer north and east values to
latitude and longitude that would be a better dataset.